+++ plugins/genai/backend/src/tools/index.ts
@@ -10,8 +10,3 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-export * from './catalogEntityTool';
-export * from './catalogSearchTool';
-export * from './techDocsSearchTool';
-export * from './techDocsReadTool';