+++ plugins/codebuild/backend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.1](https://github.com/awslabs/backstage-plugins-for-aws/compare/@aws/aws-codebuild-plugin-for-backstage-backend@0.9.0...@aws/aws-codebuild-plugin-for-backstage-backend@0.9.1) (2025-11-07)
+
+**Note:** Version bump only for package @aws/aws-codebuild-plugin-for-backstage-backend
+
+
+
+
+
 # [0.9.0](https://github.com/awslabs/backstage-plugins-for-aws/compare/@aws/aws-codebuild-plugin-for-backstage-backend@0.8.0...@aws/aws-codebuild-plugin-for-backstage-backend@0.9.0) (2025-09-19)