+++ plugins/ecr/frontend/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.1](https://github.com/awslabs/backstage-plugins-for-aws/compare/@aws/amazon-ecr-plugin-for-backstage@0.2.0...@aws/amazon-ecr-plugin-for-backstage@0.2.1) (2025-11-07)
+
+**Note:** Version bump only for package @aws/amazon-ecr-plugin-for-backstage
+
+
+
+
+
 # [0.2.0](https://github.com/awslabs/backstage-plugins-for-aws/compare/@aws/amazon-ecr-plugin-for-backstage@0.1.2...@aws/amazon-ecr-plugin-for-backstage@0.2.0) (2025-09-19)